added data to add-dog.js, move images folder to client, fixed errors in app.js
