+++ client/src/App.js
@@ -1,40 +1,40 @@
 import React from 'react';
-import Navbar from './compoments/Navigation'
-import Footer from './compoments/Footer'
-import Login from './compoments/Login'
-import Signup from './compoments/Signup'
+import Navbar from './components/Navigation'
+import Footer from './components/Footer'
+import Login from './components/Login'
+import Signup from './components/Signup'
 import UserProfile from './pages/User-Profile'
 import Matches from './pages/Search-Dogs'
 import Donation from './pages/Donation'
-import Chat from './compoments/Chat'
-import AddDog from './compoments/Add-dog'
+import Chat from './components/Chat'
+import AddDog from './components/Add-dog'
 import Home from './pages/Home'
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 
 
 
-function App() {
-  return (
-    <div className="App">
-       <BrowserRouter> 
-              <Navbar/>
-              <Routes>
-                <Route path = "/" element = {<Home/>}/>
-                <Route path = "/login" element = {<Login/>}/>
-                <Route path = "/signup" element = {<Signup/>}/> 
-                <Route path = "/profile" element = {<UserProfile/>}></Route>
-                <Route path = "/matches" element = {<Matches/>}/>
-                <Route path = "/donation" element = {<Donation/>}/>
-                <Route path = "/chat" element = {<Chat/>}/>
-                <Route path = "/add-dog" element = {<AddDog/>}/>
-              </Routes>
-              <Footer/>
-        </BrowserRouter>
-    </div>
-  )};
+// function App() {
+//   return (
+//     <div className="App">
+//        <BrowserRouter> 
+//               <Navbar/>
+//               <Routes>
+//                 <Route path = "/" element = {<Home/>}/>
+//                 <Route path = "/login" element = {<Login/>}/>
+//                 <Route path = "/signup" element = {<Signup/>}/> 
+//                 <Route path = "/profile" element = {<UserProfile/>}></Route>
+//                 <Route path = "/matches" element = {<Matches/>}/>
+//                 <Route path = "/donation" element = {<Donation/>}/>
+//                 <Route path = "/chat" element = {<Chat/>}/>
+//                 <Route path = "/add-dog" element = {<AddDog/>}/>
+//               </Routes>
+//               <Footer/>
+//         </BrowserRouter>
+//     </div>
+//   )};
 
 
-import React from "react";
+// import React from "react";
 import {
   ApolloClient,
   InMemoryCache,
@@ -45,14 +45,14 @@ import { setContext } from "@apollo/client/link/context";
 // import Cloudinary from './pages/Cloudinary';
 // import Donation from './pages/Donation';
 // import Home from './components/Home';
-import Navigation from './compoments/Navigation';
+import Navigation from './components/Navigation';
 import SearchDogs from './pages/Search-Dogs';
 // import UserProfile from './components/User-Profile';
 
 import {
   Outlet
 } from "react-router-dom";
-import Chat from "./compoments/Chat";
+// import Chat from "./compoments/Chat";
 
 // Construct our main GraphQL API endpoint
 const httpLink = createHttpLink({
