--- client/src/compoments/Signup.js
@@ -1,9 +1,0 @@
-//Test
-
-function SearchDogs() {
-    return (
-        <p>hi</p>
-    )
-}
-
-export default SearchDogs;
