+++ client/src/pages/Home.js
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
-import backgoundImg from '../../../Images/bg-img.jpg';
+import backgoundImg from '../Images/bg-img.jpg';
 import { Container} from 'react-bootstrap';
-import Header from '../components/Header';
+// import Header from '../components/Header';
 import Footer from '../components/Footer';
 import Navigation from '../components/Navigation';
